+++ lib/rails_audit_log/engine.rb
@@ -17,6 +17,8 @@ module RailsAuditLog
 
     initializer "rails_audit_log.assets" do |app|
       if app.config.respond_to?(:assets)
+        app.config.assets.paths << root.join("app/assets/stylesheets")
+        app.config.assets.paths << root.join("app/assets/images")
         app.config.assets.paths << root.join("app/javascript")
       end
     end